Add ModelSelector and responder handling in chat
Three new files were added mainly for handling the models and responders of a chat. This includes changes for ModelSelector, which allows the user to switch between various models. There's also a major refactor in the components and entities to accommodate the new model and responder handling system. Refactoring was also done in certain chat features to streamline the process.