Refactor agent selection (#16)
* Add small text class

* Add agent selector

* Overhaul responder concept

* Linting

* First draft agent selection

* Fix bug with model selection

* Remove autoprompter from template

* Add ModelSelector and responder handling in chat

Three new files were added mainly for handling the models and responders of a chat. This includes changes for ModelSelector, which allows the user to switch between various models. There's also a major refactor in the components and entities to accommodate the new model and responder handling system. Refactoring was also done in certain chat features to streamline the process.

* Fix bug with selection

* Silence react warning

* Implement refactor agent selection

* Rename to Select Responder Type

This is a group of forms

* Add agent and organization responder types

The responder type has been enhanced to include options for agents and organizations. Additional responder components for the new types were created. The configuration for OpenAI has also been added. Several new files related to these updates have been added while unnecessary syntax and lines have been removed in several existing files as part of code cleanup.

* Ignore custom properties

* Add header

* Icons

* Dry code

Add Page container

* Styling changes

* Fix bug with duplicate ids

* Fix typo

* Style icons

* Persist config

* Implicit any

---------

Co-authored-by: mgarber93 <mgarber93@gmail.com>