Update button labels and imports in form components
The labels of the submit buttons in the open-api-form and api-form components have been updated from "Submit" to "Save". In addition, the button import in api-form component has been changed from 'react-bootstrap' to our custom styled button. The position of "Api swagger" label and control in api-form has also been moved to the top for better visibility and user experience.