Reorganization (#39)
* Refactor file name

* Parameterize model

* Wip

* Reorganize files

So much neater

* I don't like gradients as much

* Restyle right border

* Lets try dashed

* Restyle header

* Refactor approval

* Return type is void

* Executor uses gpt 4

* Remove unused import

* Remove function

* Update UI styling in responder-type and aside modules

A styled div was introduced in the 'responder-type' module to enhance component arrangement. The Flexbox layout strategy was adopted and a gap was added to create more space between components. Additionally, a minor reduction in padding was made in the 'aside' module.

* Tweak animation timings

* Remove responder arg

Why specify if we're mapping later

* Remove looping

* Remove empty test

* Move snapshot file

---------

Co-authored-by: mgarber93 <mgarber93@gmail.com>