Changes (#53)
* More thoughts

* Update typescript

* Sketch continue idea

* Add 'createAndPrompt' method in AgentFactory class

The 'createAndPrompt' method has been added to the AgentFactory class which is used to create an agent and prompt it with the conversation content. Some other class adjustments have been done including the appropriate importation of dependencies in the SwaggerGptPlanProgressor class.

* WIP swagger gpt

* Simplify models

* Restyle

* Remove unneeded div

* Refactor UI styling and code organization

Several components have been modified to improve the UI and make the code more readable. These changes include the removal of redundant, unused CSS classes and variables, adjustment to the color and border of various elements, and reorganization of code imports. Moreover, some logic was simplified in message rendering, allowing for cleaner, more straightforward code.